get correct trash addressable area
